rutas publicas y privadas
